backend/requirements.txt
fastapi
youtube-transcript-api
python-dotenv
